Use Bitso fork of sqlcheck
There's a [PR open][1] to fix an [issue][2] in `sqlcheck` that causes it to break (actually hang
indefinitely) on inputs with very long lines (4096 chars or longer).

It _appears_ that `sqlcheck` is no longer being maintained, so we've forked it and released a new
version [`1.2.1` release][3] containing the fix, which we can then use in our GitHub Actions
workflows.

[1]: https://github.com/jarulraj/sqlcheck/pull/38
[2]: https://github.com/jarulraj/sqlcheck/issues/32
[3]: https://github.com/bitsoex/sqlcheck/releases/tag/1.2.1